# apps/docs/lib/engine-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "meter engine API",
    "description": "The metering engine's HTTP surface — money-truth and usage. The source of truth for engine SDK codegen.",
    "contact": {
      "name": "Josh Galbreath"
    },
    "license": {
      "name": "AGPL-3.0-only"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/v1/accounts/{id}/balance": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "`GET /v1/accounts/{id}/balance`",
        "operationId": "balance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account id (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account balance (settled / held)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Balance"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/reservations": {
      "post": {
        "tags": [
          "reservations"
        ],
        "summary": "`POST /v1/reservations`",
        "operationId": "reserve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Reservation outcome (allowed or denied)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReserveOutcome"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/usage": {
      "post": {
        "tags": [
          "usage"
        ],
        "summary": "`POST /v1/usage`",
        "operationId": "meter_usage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeterUsageBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Priced, recorded, and charged; returns credits + balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterUsageResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/usage/reservations/{id}/settle": {
      "post": {
        "tags": [
          "usage"
        ],
        "summary": "`POST /v1/usage/reservations/{id}/settle` — price the actual usage against a catalog model and",
        "description": "settle the reservation. Idempotent on the reservation id.",
        "operationId": "settle_usage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Reservation id (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettleUsageBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Credits charged + balance after",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettleUsageResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/usage/reserve": {
      "post": {
        "tags": [
          "usage"
        ],
        "summary": "`POST /v1/usage/reserve` — price a worst-case estimate against a catalog model and place a hold.",
        "description": "The engine computes the credits (ADR 0001). Returns the reserve outcome plus the reserved credits.",
        "operationId": "reserve_usage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveUsageBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Reserve outcome plus the reserved credits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReserveUsageResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Balance": {
        "type": "object",
        "description": "A point-in-time balance: settled credits, and credits currently held by open reservations.",
        "required": [
          "settled",
          "held"
        ],
        "properties": {
          "held": {
            "type": "string",
            "description": "Credits currently locked by open reservations, as an exact decimal string."
          },
          "settled": {
            "type": "string",
            "description": "Settled credit balance, as an exact decimal string."
          }
        }
      },
...
      "MeterUsageBody": {
        "type": "object",
        "description": "`POST /v1/usage` — price token usage via the catalog, record the event, and charge credits.",
        "required": [
          "org_id",
          "account",
          "model",
          "idempotency_key"
        ],
        "properties": {
          "account": {
            "type": "string",
            "format": "uuid"
          },
          "idempotency_key": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "rate_card_id": {
            "type": "string",
            "description": "Optional synced rate-card id to price against; defaults to the catalog card for `model`.",
            "nullable": true
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "usage": {
            "$ref": "#/components/schemas/UsageDimensions"
          }
        }
      },
      "MeterUsageResult": {
        "type": "object",
        "description": "`POST /v1/usage` result: the priced amounts, the recorded event id, and the resulting balance.\nCredit/USD amounts are exact decimal strings.",
        "required": [
          "credits",
          "cogs_usd",
          "customer_price_usd",
          "event_id",
          "charged",
          "settled",
          "available"
        ],
        "properties": {
          "available": {
            "type": "string"
          },
          "charged": {
            "type": "boolean"
          },
          "cogs_usd": {
            "type": "string"
          },
          "credits": {
            "type": "string"
          },
          "customer_price_usd": {
            "type": "string"
          },
          "event_id": {
            "type": "string",
            "format": "uuid"
          },
          "settled": {
            "type": "string"
          }
        }
      },
...
      "ReserveBody": {
        "type": "object",
        "description": "`POST /v1/reservations`",
        "required": [
          "account",
          "reservation_id",
          "amount",
          "limit"
        ],
        "properties": {
          "account": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Optional hold expiry (RFC3339); an open hold past it is released by the sweep.",
            "nullable": true
          },
          "limit": {
            "type": "string",
            "description": "Limit class: `hard` (fail-closed) or `soft` (fail-open)."
          },
          "reservation_id": {
            "type": "string",
            "format": "uuid"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "description": "Optional agent run this hold belongs to; lets `POST /v1/runs/{id}/void` reverse it.",
            "nullable": true
          }
        }
      },
      "ReserveOutcome": {
        "oneOf": [
          {
            "type": "object",
            "description": "The hold was placed (or already existed); the caller may proceed to spend.",
            "required": [
              "reservation",
              "outcome"
            ],
            "properties": {
              "outcome": {
                "type": "string",
                "enum": [
                  "allowed"
                ]
              },
              "reservation": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "type": "object",
            "description": "A HARD limit refused the spend; the call must not proceed.",
            "required": [
              "available",
              "requested",
              "outcome"
            ],
            "properties": {
              "available": {
                "type": "string"
              },
              "outcome": {
                "type": "string",
                "enum": [
                  "denied"
                ]
              },
              "requested": {
                "type": "string"
              }
            }
          }
        ],
        "description": "The result of a [`ReserveRequest`].",
        "discriminator": {
          "propertyName": "outcome"
        }
      },
      "ReserveUsageBody": {
        "type": "object",
        "description": "`POST /v1/usage/reserve` — price a worst-case usage estimate against a catalog model and place a\nhold before the call. The engine computes the credits (ADR 0001), not the caller.",
        "required": [
          "account",
          "reservation_id",
          "model",
          "limit"
        ],
        "properties": {
          "account": {
            "type": "string",
            "format": "uuid"
          },
          "estimate": {
            "$ref": "#/components/schemas/UsageDimensions"
          },
          "limit": {
            "type": "string",
            "description": "Limit class: `hard` (fail-closed) or `soft` (fail-open)."
          },
          "model": {
            "type": "string"
          },
          "rate_card_id": {
            "type": "string",
            "description": "Optional synced rate-card id to price against; defaults to the catalog card for `model`.",
            "nullable": true
          },
          "reservation_id": {
            "type": "string",
            "format": "uuid"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "description": "Optional agent run this hold belongs to; lets `POST /v1/runs/{id}/void` reverse it.",
            "nullable": true
          }
        }
      },
      "ReserveUsageResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ReserveOutcome"
          },
          {
            "type": "object",
            "required": [
              "reserved_credits"
            ],
            "properties": {
              "reserved_credits": {
                "type": "string",
                "description": "Credits the engine priced and reserved, as an exact decimal string."
              }
            }
          }
        ],
        "description": "`POST /v1/usage/reserve` result: the reserve outcome (flattened) plus the engine-computed credits\nthat were held."
      },
...
      "SettleUsageBody": {
        "type": "object",
        "description": "`POST /v1/usage/reservations/{id}/settle` — price the actual usage against a catalog model and\nsettle the reservation. Idempotent on the reservation id.",
        "required": [
          "model"
        ],
        "properties": {
          "actual": {
            "$ref": "#/components/schemas/UsageDimensions"
          },
          "model": {
            "type": "string"
          },
          "rate_card_id": {
            "type": "string",
            "description": "Optional synced rate-card id to price against; defaults to the catalog card for `model`.",
            "nullable": true
          }
        }
      },
      "SettleUsageResult": {
        "type": "object",
        "description": "`POST /v1/usage/reservations/{id}/settle` result.",
        "required": [
          "credits_charged",
          "balance_after"
        ],
        "properties": {
          "balance_after": {
            "type": "string"
          },
          "credits_charged": {
            "type": "string"
          }
        }
      },
...
      "UsageDimensions": {
        "type": "object",
        "description": "Token counts for a metered usage event.",
        "properties": {
          "cache_read": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "cache_write": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "input_uncached": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "output": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "reasoning": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      }
    }
  },
...
}